Trimming utils to only host functions that are used in multiple files.
